LayoutTests/fast/loader/stateobjects/document-destroyed-navigate-back-with-fragment-scroll.html

<html>
<head>
<script>

if (window.layoutTestController) {
    if (!sessionStorage.stage)
        layoutTestController.clearBackForwardList();
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function log(txt)
{
    document.getElementById("logger").innerText += txt + "\n";
    // alert the messages also so DumpRenderTree can capture and log messages across multiple documents.
    alert(txt);
}

function endTest(msg)
{
    log(msg);
    sessionStorage.clear();
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

function lastPathComponent()
{
    return window.location.href.split('/').pop();
}

function runSecondStageOfTest()
{
    log("History length is " + history.length);
    log("Last path component of location is " + lastPathComponent());
    sessionStorage.stage = 3;
    history.back();
}

function runThirdStageOfTest()
{
    log("History length is " + history.length);
    log("Last path component of location is " + lastPathComponent());
    endTest("Test completed");
}

function loaded()
{
    if (sessionStorage.stage) {
        if (sessionStorage.stage == 2)
            runSecondStageOfTest();
        else if (sessionStorage.stage == 3)
            endTest("Shouldn't reach this case when doing fragment scrolls");
        else
            endTest("Unexpected stage value");
    } else
        runFirstStageOfTest();
}

function runFirstStageOfTest()
{   
    history.replaceState("FirstEntry", null, "#FirstEntry");
    history.pushState("SecondEntry", null, "#SecondEntry");
    
    log("History length is " + history.length);
    history.back();
}

function statePopped()
{
    log("State popped - " + event.state + " (type " + typeof event.state + ")");
    if (event.state == "FirstEntry") {
        history.replaceState("FirstEntryShouldNeverBeReactivated", null, "#FirstEntryShouldNeverBeReactivated");
        history.forward();
    } else if (event.state == "SecondEntry") {
        history.replaceState("SecondEntryShouldNeverBeReactivated", null, "#SecondEntryShouldNeverBeReactivated");
        window.location = "resources/navigate-back.html";
    } else
        endTest("Unexpected state popped - " + event.state);
}

function hashChanged()
{
    log("hashChanged - Last path component of location is " + lastPathComponent());

    if (window.location.hash == "#FirstEntryShouldNeverBeReactivated")
        endTest("Test completed");
}

</script>
<body onload="loaded();" onpopstate="statePopped();" onhashchange="hashChanged();" onunload="/* disable page cache */">
<pre>
This test:
-Builds up a list of state object entries with fragment URL.
-Navigates through them to verify that the popstate and hashchanged events are fired.
-Navigates away to a new document, with the old document being destroyed.
-Navigates back to the state object entries and verifies the popstate event is not fired.
</pre><br>
<pre id="logger"></pre>
</body>
</html>
